fastapi-service/requirements.txt
fastapi
uvicorn
pydantic
gitpython
faiss-cpu
sentence-transformers
openai
python-dotenv
langchain
langchain-google-genai
google-generativeai
chromadb
python-multipart
